// dashboard/src/api/models/numberingSystemsListAppliesTo.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Invoicence API
 * Description
 * OpenAPI spec version: 1.0.0
 */

export type NumberingSystemsListAppliesTo =
  (typeof NumberingSystemsListAppliesTo)[keyof typeof NumberingSystemsListAppliesTo];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const NumberingSystemsListAppliesTo = {
  credit_note: "credit_note",
  invoice: "invoice",
  quote: "quote",
} as const;