// dashboard/src/api/models/taxRatesListStatus.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Invoicence API
 * Description
 * OpenAPI spec version: 1.0.0
 */

export type TaxRatesListStatus =
  (typeof TaxRatesListStatus)[keyof typeof TaxRatesListStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TaxRatesListStatus = {
  active: "active",
  archived: "archived",
} as const;